Fixes MSVC warnings

+++ tools/RunGen.cpp
@@ -901,8 +901,8 @@ int main(int argc, char **argv) {
     bool track_memory = false;
     bool describe = false;
     double benchmark_min_time = BenchmarkConfig().min_time;
-    int benchmark_min_iters = BenchmarkConfig().min_iters;
-    int benchmark_max_iters = BenchmarkConfig().max_iters;
+    int benchmark_min_iters = static_cast<int>(BenchmarkConfig().min_iters);
+    int benchmark_max_iters = static_cast<int>(BenchmarkConfig().max_iters);
     for (int i = 1; i < argc; ++i) {
         if (argv[i][0] == '-') {
             const char *p = argv[i] + 1; // skip -
